Moved Android code to shared
