+++ androidApp/src/main/java/fr/paug/androidmakers/ui/MainActivity.kt
@@ -16,6 +16,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.core.view.WindowCompat
 import androidx.lifecycle.lifecycleScope
+import com.androidmakers.ui.about.AboutActions
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
@@ -30,9 +31,7 @@ import fr.androidmakers.domain.model.Partner
 import fr.androidmakers.store.firebase.toUser
 import fr.paug.androidmakers.R
 import fr.paug.androidmakers.ui.components.MainLayout
-import fr.paug.androidmakers.ui.components.about.AboutActions
 import fr.paug.androidmakers.ui.theme.AndroidMakersTheme
-import fr.paug.androidmakers.util.CustomTabUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
