+++ androidApp/src/main/java/fr/paug/androidmakers/ui/components/session/SessionDetailLayout.kt
@@ -60,17 +60,17 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
+import com.androidmakers.ui.common.LoadingLayout
+import com.androidmakers.ui.model.Lce
 import fr.androidmakers.domain.model.Room
 import fr.androidmakers.domain.model.Session
 import fr.androidmakers.domain.model.Speaker
 import fr.paug.androidmakers.AndroidMakersApplication
 import fr.paug.androidmakers.R
 import fr.paug.androidmakers.ui.MR
-import fr.paug.androidmakers.ui.components.LoadingLayout
 import fr.paug.androidmakers.ui.theme.AMColor
 import fr.paug.androidmakers.ui.util.discardHtmlTags
 import fr.paug.androidmakers.ui.util.stringResource
-import fr.paug.androidmakers.ui.viewmodel.Lce
 import fr.paug.androidmakers.util.EmojiUtils
 import io.openfeedback.android.components.SessionFeedbackContainer
 import separatorColor
