+++ androidApp/src/main/java/fr/paug/androidmakers/ui/model/uimodel.kt
@@ -16,12 +16,3 @@ class UISession(
 ) {
   class Speaker(val name: String)
 }
-
-class UIVenue(
-    val imageUrl: String,
-    val name: String,
-    val address: String?,
-    val coordinates: String?,
-    val descriptionEn: String,
-    val descriptionFr: String,
-)
