+++ shared/build.gradle.kts
@@ -27,6 +27,9 @@ kotlin {
   sourceSets {
     commonMain.dependencies {
       api(libs.moko.resources)
+
+      // Temporary
+      api(libs.precompose.koin)
       api(project(":shared:ui"))
       api(project(":shared:domain"))
       api(project(":shared:di"))
