+++ shared/ui/src/commonMain/kotlin/com/androidmakers/di/ViewModelModule.kt
@@ -1,0 +1,14 @@
+package com.androidmakers.di
+
+import com.androidmakers.ui.speakers.SpeakerListViewModel
+import com.androidmakers.ui.sponsors.SponsorsViewModel
+import com.androidmakers.ui.venue.VenueViewModel
+import org.koin.dsl.module
+
+val viewModelModule = module {
+  factory { SpeakerListViewModel(get()) }
+  factory { SponsorsViewModel(get()) }
+  factory { VenueViewModel(get(), get()) }
+
+
+}
