+++ shared/ui/src/commonMain/kotlin/com/androidmakers/ui/model/UIVenue.kt
@@ -1,0 +1,10 @@
+package com.androidmakers.ui.model
+
+class UIVenue(
+    val imageUrl: String,
+    val name: String,
+    val address: String?,
+    val coordinates: String?,
+    val descriptionEn: String,
+    val descriptionFr: String,
+)
