+++ shared/ui/src/commonMain/kotlin/com/androidmakers/ui/sponsors/SponsorsViewModel.kt
@@ -1,0 +1,19 @@
+package com.androidmakers.ui.sponsors
+
+import at.asitplus.KmmResult
+import com.androidmakers.ui.common.LceViewModel
+import fr.androidmakers.domain.interactor.GetPartnersUseCase
+import fr.androidmakers.domain.model.PartnerGroup
+import kotlinx.coroutines.flow.Flow
+
+class SponsorsViewModel(
+    private val getPartnersUseCase: GetPartnersUseCase
+) : LceViewModel<List<PartnerGroup>>() {
+  override fun produce(): Flow<KmmResult<List<PartnerGroup>>> {
+    return getPartnersUseCase()
+  }
+
+  init {
+    launch(false)
+  }
+}
