+++ shared/ui/src/commonMain/kotlin/com/androidmakers/ui/venue/FloorPlan.kt
@@ -1,4 +1,4 @@
-package fr.paug.androidmakers.ui.components
+package com.androidmakers.ui.venue
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
@@ -6,14 +6,14 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import fr.paug.androidmakers.R
+import dev.icerock.moko.resources.compose.painterResource
+import fr.paug.androidmakers.ui.MR
 
 @Composable
 fun FloorPlan() {
   Box(modifier = Modifier.verticalScroll(rememberScrollState())) {
     Image(
-        painter = painterResource(id = R.drawable.beffroi1800px),
+        painter = painterResource(MR.images.beffroi),
         contentDescription = "floor plan"
     )
   }
