+++ shared/ui/src/commonMain/kotlin/com/androidmakers/ui/venue/VenueViewModel.kt
@@ -1,8 +1,9 @@
-package fr.paug.androidmakers.ui.components.venue
+package com.androidmakers.ui.venue
 
-import androidx.lifecycle.ViewModel
 import fr.androidmakers.domain.interactor.GetAfterpartyVenueUseCase
 import fr.androidmakers.domain.interactor.GetConferenceVenueUseCase
+import moe.tlaster.precompose.viewmodel.ViewModel
+
 
 class VenueViewModel(
     val getConferenceVenueUseCase: GetConferenceVenueUseCase,
