+++ shared/ui/src/commonMain/moko-resources/fr/strings.xml
@@ -2,7 +2,7 @@
 <resources>
     <string name="app_name">Android Makers by droidcon</string>
 
-    <string name="version">Version %s (%d)</string>
+    <string name="version">Version %s (%s)</string>
 
     <string name="agenda">Agenda</string>
     <string name="venue">Lieu</string>
