+++ shared/ui/src/commonMain/moko-resources/images/beffroi@1x.png
