+++ shared/ui/src/iosMain/kotlin/com/androidmakers/ui/speakers/SpeakerListViewController.kt
@@ -1,25 +1,18 @@
 package com.androidmakers.ui.speakers
 
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.window.ComposeUIViewController
-import at.asitplus.KmmResult
-import com.androidmakers.ui.model.toLce
 import com.androidmakers.ui.theme.AndroidMakersTheme
-import fr.androidmakers.di.DepContainer
+import moe.tlaster.precompose.koin.koinViewModel
 import platform.UIKit.UIViewController
 
-fun SpeakerListViewController(): UIViewController =
+fun SpeakerListViewController(onSpeakerClick: (String) -> Unit): UIViewController =
     ComposeUIViewController {
-      val depContainter = DepContainer()
-      val speakersRepository = depContainter.speakersRepository
-      val speakers by speakersRepository.getSpeakers().collectAsState(KmmResult.success(emptyList()))
-      val content = speakers.toLce()
 
       AndroidMakersTheme {
+        val viewModel = koinViewModel(vmClass = SpeakerListViewModel::class)
         SpeakerScreen(
-            content = content,
-            navigateToSpeakerDetails = {}
+            viewModel = viewModel,
+            navigateToSpeakerDetails = onSpeakerClick
         )
       }
     }
